requirements.txt
Django==3.2
whitenoise==6.0.0
gunicorn
django-cors-headers
djangorestframework
openai